# Conflicts:
#	Project/Assets/Scripts/GUI/GameLoadingController.cs
#	Project/Assets/Sparta/Config/Resources/Installers/GlobalDependencyConfigurer.asset

+++ Project/Assets/Sparta/Extensions/Social/SocialFramework/Chat/BaseChatMessage.cs
@@ -15,7 +15,7 @@
 
         public int RankInAlliance { get; set; }
 
-        public long PlayerLevel { get; set; }
+        public int PlayerLevel { get; set; }
 
         public int AllianceAvatarId { get; set; }
